pyproject.toml
[project]
name = "PyALM"
version = "0.1.0"
authors = [
  { name="Finn Schwall", email="finn.schwall@isob.fraunhofer.de" },
]
description = "Python abstraction layer for large language models"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache 2",
    "Operating System :: OS Independent",
]

dependencies = [
    "psutil",
    "tqdm",
    "numpy",
    "pyyaml",
    "pynvml",
    "openai",
    "tiktoken~=0.5.1",
]
[project.optional-dependencies]
openai = [

]
aleph_alpha = ["aleph_alpha_client"]
llama = ["llama-cpp-python"]
full = ["tiktoken",
     "openai","aleph_alpha_client","llama-cpp-python"]



[project.urls]
"Homepage" = "https://github.com/finnschwall/PyALM"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
